requirements.txt
flask
bulma
pyyaml
folium
gpxpy